Fix messaging around the InvalidURI exception
The default behavior of InvalidURI would display the config_uri, but if you specify a message, that config_uri is not displayed, making troubleshooting hard.  This adds `config_uri` to the exception message so you can tell what is broken.  In my case it was
`pshell -p bpython my.ini`  which needed to be `pshell my.ini -p bpython`
The exception message before this change didn't tell you what you needed to know (that `-p` was the `config_uri`)